[IMP] event: send mail on event stage change
Purpose
=======
Be able to send event mail when we move an event to another stage.

e.g.: Sending "Sorry for canceling the event" when we move the event
in the stage "Canceled".

Specifications
==============
If we specify "On stage update" on the event mail, the mail is sent
scheduled when we move the event into the specified stage.

If we specify a interval, the event mail will be sent "X times" after
moving the event into the specified stage.

Reset the `trigger_stage_id` if the interval type is not "stage_update".

Task 2127615
PR #45200
See odoo/upgrade/pull/858