[IMP]documents: maximal size detection
PURPOSE

Currently, when we add a file too large, we wait some times before we are
notified by Odoo "413 Request Entity too large"

SPECIFICATIONS

- Catch correctly the 403 error to display a user friendly message instead of
the technical error that doesn't mean something for common humans. But this will
be displayed when the upload already started.
- Add a ir.config.parameter (System Parameter, see screenshot), unset by default
But let's say that the limit on the instance is set to 500MB and if the
administrator of the database wants to set the limit to 50MB, it could be
possible from the interface, and in that case, it would be possible to display a
modal before the upload starts to notify the user that his file is too big.
- Set the limit correctly for saas databases, as we are in charge of the hosting
 (and thus the nginx configuration).