[IMP] event_*: clean ACL and add a new group "Registration Desk"
Purpose
=======
Clean the ACLs related to the Event application.

Add a new group to manage the registration in the entrance of an event.
This group should not be able to "break" things in Event but should be
able to create and manage the registrations.

Specifications
==============
Now, there are 3 event groups
- The Registration Desk User, who can manage the registrations and
  that's it.
- The Event User, who can create event, sponsor, ticket... His role is
  to manage the events.
- The Event Administrator, who can create event type, sponsor type,
  ticket type... His role is to manage the way we use the event
  application.

Each group implies all previous groups.

Task 2204364