[FIX-IMP] im_livechat: aggregate livechat statistics by livechat_channel
Before this commit, livechat statistics was aggregated by mail.channel
technical name, which is not really a usefull infirmation, and could lead
to a huge amount of "bars" in graph, so huge that it becomes unreadable.

After this commit, we aggregate by livechat channel. Also, pivot table is
divided by session date (by month), in order to make it more readable.

Task ID: 2301261