[FIX] (test_)mail: edit subtype subscriptions is not saved in some case
Before this commit:

Currently, the subtype checked selection is not saved when the user adds and
removes subtypes.
For example: click on the pencil icon to edit the subscription of
the follower. Normally by default "Discussions" will be the only one selected.
Check notes and save its working. Now click again on the pencil icon and uncheck
notes and check activity and save. Here in this case notes have been unchecked
but activity is not checked.

The problem only occurs when you unchecked subtypes and checked other ones.

After this commit:

The subtype checked selection will be saved when the user add and remove
subtypes.

LINKS

PR https://github.com/odoo/odoo/pull/56775
Task-2241688