[IMP] mass_mailing_sms: Add KPI Recap for SMS
Added SMS KPIs analysis report. This report will be sent by email 24 hours after the campaign sms are sent.
Modification of the digest_mail_layout to be able to use it to generate the KPI report for Email or SMS mass mailing campaigns.
The goal is to have a single template that can be used for both SMS and email campaign.
The changes in mailing_mailing for email and sms are intended to make this possible.

Task-2274770