[IMP] hr: add a model for work locations
The work location of an employee was previously a fields.Char. That was changed to a Many2one field to a the new model called work.location.
That had some consequences in other modules that had to be slightly adjusted.
The modules impacted were hr, hr_appraisal, hr_payroll.

Task id: 2335646