[IMP] web: replace create_text node options with attribute
before this commit: create_text was passed in node options and due to which
it was not parsed by translate.py, pass create_text as attribute on field
so that translate.py parse it and it is translated.

after this commit: create_text will be passed as field attribute instead of
node options.

task-1923433