[FIX] account: remove tip on onboarding tour
When the user is asked to fill in the label on the invoice, the tooltip
tells him that the label can be auto-filled by configuring its products.
This leads the user to configure products. But then, as he doesn't fill
in the label, the tour is "broken".

In addition, on the PDF invoice, the line about payment communication
and the block with taxes are too close. Create blank space between them.