[ADD] l10n_demo: make demo data for all localization
Task 2124952

Extract the demo data of l10n_generic_coa in a new module so that the
demo data is installed for all localization.
This module is installed after the instanciation of chart templates for
the company if demo data is needed for `account`

To find the correct accounts, a helper method has been added to
`account.account`. It will look for an account with the correct xml_id
(which is the one of the l10n_generic_coa, that we have to put on the
other l10n modules too, in the future). If it didn't find the xml_id, it
will look for an account with the correct type. If it still didn't find
one, just take any account.