[FIX] mail: Update the kanban record on activity creation/updation
Currently, scheduling the activity from the document chatter does not
appear on the kanban record on activity button.
and the 'onchange' action receives the ID of the activity that was
marked as done and thus the 'Missing record' error

So in this commit, on creation and done of activity reload th  view to
update the record on kanban card.

TaskID: 2362964