[FIX] web_editor: DOM is not cleaned before saving custom snippets
Previously, when saving a custom snippet, the outerHTML of the snippet
was naively read directly and saved as a new custom snippet.
Unfortunately this completely bypasses all cleanup mechanisms that we
have before saving, such as stopping the public widgets, or creating
attachments for cropped or optimized images. This would result in weird
bugs, where public widgets would appear twice when dropping a custom
snippet containing one, as well as the creating a new attachment for any
image that was optimized or cropped before saving the custom snippet
every time it was dropped in the page again.

This commit fixes that by adding a request_clean event to the wysiwyg
and factoring out the cleaning logic into a separate function, it is
then called by the snippetsMenu before creating a new custom snippet.

task-2341612

X-original-commit: d649232e49a8bca55ba8684a605fd34068138cf6