[IMP] hr_expense: create report from list view should open in edit mode
Currently, when user click on 'create report' button form list view
header, system create report but on create report page the summary field
is required and it's empty so to fill that value the report will be
opened in edit mode.
Also same action coming to the two places
1) in top action menu
2) on tree header

So in this commit,when expense report is open from the tree view is
will be in edit mode and removed 'create report' menu from top
action bar.

TaskID: 2359834