[IMP] crm: allow user to cancel updation of probabilities from settings
Right now, we have a button in the CRM settings called 'Update Probabilities'
clicking on which, probabilities of the leads created after certain date
will be updated. Problem with this flow is, user can click on the button by
mistake, and there's no going back.

This commit improves the behavior by opening a wizard on click of the button
instead of directy updating the probabilities, thus allowing user to cancel
the action in case button was clicked by mistake.

TaskID - 2355562