[FIX] web_editor: Do not prompt when leaving in backend
Before this commit: when trying to exit a view in which a wysiwyg editor
was instanciated: an unwanted prompt popped up. Furthermore, if the
editor was inside a field_html it was flagged as dirty.

Now: the exiting prompt only shows up in translation or website mode and
the field is flagge as dirty only when inputting acutal changes.

Task 2371592