[FIX] stock_account: currency in forecast report
Before this commit, the SVL domain wasn't correctly set. So we got back
SVL for all companies, which can lead to an error if the companies use
different currencies.

How to reproduce:
  - Create a new company with a different currency than the existing one
  - For a product, define a cost and add some qty. on hand for the two
    companies;
  - Display the Forecast Report for this product -> traceback.

To fix that, we use `product.product` `value_svl` field instead of
recomputing the value from the SVL and we do that only for the current
selected warehouse's company, so we deal with only one currency at the
time.