[IMP] website_crm: Catch partner info on contact form
Reduce friction submitting a CRM contact us form while linking the lead to the portal contact.
This facilitates the flow generating leads from the contact us form.

The contact us form is pre-filled with data from the partner linked with logged in website visitor using t-att-data-values to set
default values in xml. If some data is set in request.params, then it is used in priority as default values (previous behaviour implemented).
The partner_id of the lead created from form is set to the one of the logged visitor if model is crm.lead and there is a logged_visitor.

Customer : Luc Nailis (lna)
Task-ID : 2241639