[FIX] stock: display float with correct decimal
Before this commit, on the delivery slip, the quantities aren't
displayed at the same way if we display the move ones or move line ones.
In the first case, we print directly the value from a record, so the
float decimal is correclty placed when the float is converted into a
string.
In the second case, we get the value from a dict, so there is no
conversion and the float is simply translated into a string.

For example, for 2 qty.:
  - For a move it display 2.00
  - For a move line it display 2