[IMP] web: limitate the clickable area for an email widget
The 'o_text_overflow' class whlie provied on any field, is used to
have classic ellipsis (...) for very long string. However, it is
implementd the way that it occupies the availabe width for the field
tag it even if the the string is small.

The same class is also utilized in the email widget. But it has a small
side-effect due to full width occupancy. Because the email widget in
this case renderes an anchor tag, he clickable area is on the whole
availabe width instead of simply on the text. So even if user clicks
on the empty area of that field, a mail client pops-up which should
not happen (see how clickable m2o fields work on form view, for ex).

There seems to be no easy way to fix this from css only. With this
commit, we wrap the acnhor tag, within a container div tag. Here,
the overflow class will be on the container div which will do it's
job to prevent the long strings from breaking the UI, and the anchor
tag being its child, will not be the full width, thus limiting the
clickable area.

TaskId - 2345974