[FIX] mail: remove 'ratings' to smiley face
Before this commit:
For an example, when task is going to 'Done' state at that time mail is sent to
customer for feedback, answering of that mail contain smiley face plus addition
description in chatter.

After this commit:
An addition description is removed to the smiley face in message.

Links
PR https://github.com/odoo/odoo/pull/62517
Task-2373127