[REF] mail: make activity_user_type not required on ir.actions.server model
PURPOSE

Required fields are annoying, notably when dealing with tests or xml-based data
to create in concurrent modules. For example ModuleA adds a required field on
Model. ModuleB that does not depend on ModuleA tries to create a data of
Model. Currently it crashes as field is required but default is not provided if
ModuleA is not loaded in registry.

SPECIFICATIONS

Field activity_user_type on ir.actions.server model in mail is currently
required. This field is however not really required as its use is limited
to actions linked to activities. Remove the hard required and put it only
in views.

If no value is given due to some bad configuration no responsible will be
set.

Followup of odoo/odoo@730998142678dc753fc77fb95df1643d0ccb7526

LINKS

Task ID-2367655