[FIX] core: consistent environment for env.company(ies)
Before this commit, the company recordset returned by env.company/env.companies
was sometimes in sudo and sometimes not.

If 'allowed_company_ids' was specified in the context, the returned recordset wasn't sudoed.
If no company was specified through the context, company/companies was using env.user.company_id(s)
as fallback, which is always sudoed since env.user is always sudoed.

With this commit, the behavior is now consistent.