[IMP] website_slides: separate backend tags from frontend tags
In general, Backend tags are used internally, for classification or daily
process. On the other end, Frontend tags are used for filtering and menu
definitions. These two type of tags are not managed separately in eLearning
channels right now.

This commit separates the backend tags with frontend tags with help of the
color field for the channels. Along with that, it improves the following :

Frontend :
- only displays channel tags that has color set
- avoids showing void tag group / menu (without any frontend tags)

Backend :
- prevents access of the backend tags for portal/public users with a
  record rule (similar to online events)
- adds 'Tags (tag_ids)' field in the list view of 'Course Groups' that
  shows all the tags (frontend as well as backend)

Task-id: 2357942