[FIX] web_editor: fix scroll event on snippet popup
Before this commit, when scrolling a snippet popup, the overlay was
not updating its position.

So now, we use addEventListener instead of jQuery for the scroll event
on the scrolling element because we need the 'capture' parameter.
Setting this capture parameter to true allows us to take advantage of
event bubbling for events that otherwise don’t support it.

Co-authored-by: sdegueldre <sad@odoo.com>

Part of #61998

task-2312878