[IMP] project : Improve sub tasks creation and workflow
We want to ease user experience regarding sub task creation and management.
We thus enforce the values of subtask stage, project and parent task. we want
to remove the ability of a user to choose a third-party subtask project different
than the actual one. Also, we want to remove the ability to make one sub task recurrent.
In general, We want to improve and strengthen the relation parent-childs for Tasks. A sub task is dependant of its parent all along the workflow.

This commit address those needs and, especially, removes the subtask project configuration discussed in :
- https://github.com/odoo-dev/odoo/commit/66a0e5a541f7311bf7bb05278654b329ae485ea0
- https://github.com/odoo-dev/odoo/commit/b709e22fa2911b7a2ffd8efda7e00845a887a03d

Task-2388034 : Part of task-2379043