[FIX] core: Fix invalidation of computed editable field in one2many
Introduced by:
https://github.com/odoo/odoo/commit/fd50ba9fb838067fab0b920470cf79b9c94d785c#diff-7144f88ea32f36feb17ce1b8dda7dee1631f5ada34075414587df3948c6b3d1bR6148

The way the computed fields are invalidated depends of the dict order passed as parameter to the 'onchange'
method. This is causing some unexpected behavior when dealing with computed editable fields because the user
is loosing its custom value right away during the 'onchange'.

Explanation:
During the 'onchange', the one2many value is cache by record._update_cache(changed_values, validate=True).
If the field is a one2many, 'value' contains a (1, _, {...}) command for updated lines.
Due to 'validate=True', the assignations are triggering invalidation.
In case there is anything on the parent model editing any line value, the whole one2many values are returned to
the user.