[IMP] web: disable listview buttons on click and enable it once action finished
before this commit: listview buttons on row were not disabled on as soon as
it is clicked due to which user can click button multiple time and action
may trigerred multiple times.

after this commit: listview row buttons will be disabled as soon as it is
clicked and will enabled once action performed on that button is done.

task-2391298