[IMP] web: add quick edit behaviour
This commit adds the quick edit behaviour.
The quick edit allows to click on fields in readonly form view to switch
into edit mode. After switching mode, the clicked in automatically
focused. Clicking on a checkbox or its label automatically toggles it
after switching mode and clicking on a radio button auto sets its value.
Changing page or going back in the breadcrumb automatically saves the
record if we are in quick edit.

Default edit/save (with controller buttons) behaviour still exists, this
behaviour does not automatically save the record.

task 2330101