[IMP] web: handle buttons with long strings on quick create kanban card
Right now, the quick create kanban card is having three buttons, Add, Edit
and Discard. In some languages, the translation terms of these buttons are
logner, which pushes the last button into second line.

This commit replaces the 'Discard' button with trash icon, and
makes sure that all the buttons are always in single line. If the
strings are too long, they will be cropped automatically.

TaskID - 2373130