[FIX] account: Fixes the losing of company_id in account move
Currently, the company_id of an account move is related to its journal.
This means that if you empty the journal field, the company_id is also
emptied.
This is then causing a traceback when adding a new line, for example.

There is another issue in v14 where the list of possible journals is
re-computed when the company_id change, effectively blocking the user
from adding back the journal.

This change will transform the related company_id in a computed field,
making sure to not lose the company_id if the journal_id is emptied.

task id #2392004