[IMP] web_editor: execCommand automatic context
Automatically use the current context of the editor.execCommand
when a command is running (i.e. microtastks not finished).

In the future commit where we will introduce the system that
observe mutation to translate the change in the editor VDOC,
this system will become very handy because we will not need
to pass context throughout all the functions that need to execute
a command.