[FIX] crm: avoid unexpected phone warning on lead
With a recent commit[1], we fixed the issue of false warning while changing
the partner on lead. But that fix ignored the fact that the lead can have
partner and phone set with demo data. So, while viewing the form view for
such leads, the compmute method for ribbon warning considers unformatted
phone and compares it with the formatted phone of the partner, which makes
the warning appear everytime even if no change has been made.

This commit fixes the issue by formatting both lead phone and parter phone
before comparing them and thus avoiding falsy phone warning.

[1] - https://github.com/odoo/odoo/commit/d29862c1de6208f7fdb7c8104afe0d35241d130b

TaskID - 2390287