[FIX] mail: emoji popover on mobile is opening before keyboard close
Before this commit:The popover is opened before the mobile keyboard closes,
which means there will be a flicker, which means trying to click on an emoji
just as the keyboard is closing will be very confusing, or even lead to a wrong
action.

After this commit:The popover will open after the mobile keyboard closes.