[FIX] tests: kill remaining chrome headless instances
When an HttpCase goes wrong for one reason or another, a Chrome headless
browser may stay running in the background after the test session is
finished.

While it's not a problem on the runbot system as the Docker containers
are recycled, it may be a problem on other systems.

Latest use case: the requests.get is patched in a test but not
unpatched ... This made the _json_command method completely fails
leaving a Chrome headless browser running.