[FIX] sale_timesheet: add the default timesheet product in project
When the user checks the allow_billable and save, we have an validation
error because the timesheet_product_id=False.

In fact, the compute method to correctly set on the project model but
not in the definition of the field.
This commit add the compute method in the field and keep the previous
behaviour. That is, we keep this field as store and editable field.

task-2409761