[IMP] web: save multi edit listview on save button
scenario: user selects multiple records, edit it and click on discard button
and then click on cancel, so rows are still in edit mode, now user clicks on
save button then save button does not work, as multi edit works on
field_changed only.

after this commit: we store lastFieldChanged event and check in saveRecord
method if multipleRecordsSavingPromise is not available then use
lastFieldChangedEvent and call _fieldChanged with lastFieldChangedEvent.

task-2376233