[ADD] website_slides: separate backend tags from frontend tags
now in kanban view we can hide the tags while "hide for kanban"
is enabled, while creating new tags defaultly they had no colors
but now they will have colors defaultly that means "hide for kanban"
is defaultly false now.
In frontend while a user added a tag it was not shown in the frontend
because defaultly the tag was created without any color, now as tags are
created with color they will be displayed in the frontend as the tags
have color, so by this the confussion is resolved, no one can create
colorless tags from frontend.

Task-id: 2357942